Change expected values based on a discussion with @jrgm
While the original expected values were more correct in the
majority of these tests, we shouldn't impose too much logic on the
nginx rules that will be making the tests pass.

It's better to have slightly less precise HTTP error codes but to
have an easier (and more maintainable) nginx config.